Move getStaticAuthDefinitions to a different file to avoid circular deps
